Remove some warnings.

+++ Base/base.hpp
@@ -9,7 +9,7 @@ void LogTtfError(const std::string& msg);
 
 const float EPS = 0.0001f;
 
-const float TO_DEG = 180.0f / M_PI;
-const float TO_RAD = M_PI / 180.0f;
+const float TO_DEG = static_cast<float>(180.0 / M_PI);
+const float TO_RAD = static_cast<float>(M_PI / 180.0);
 
 #endif // BASE_H
